Implemented useGetProfileData hook for Profile screen.
