+++ src/screens/ProfileScreen/ProfileScreen.js
@@ -1,12 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import "./ProfileScreen.css";
-import { useParams } from "react-router-dom";
 import Header from "../../components/Header/Header";
-import { useSelector } from "react-redux";
-import axios from "axios";
 import ProfileArticleCard from "../../components/ProfileArticleCard.js/ProfileArticleCard";
 import Loader from "../../components/Loader/Loader";
-import { useLocation } from "react-router-dom";
 import { breadcrumbs } from "../../data/data";
 import AuthorProfile from "../../components/AuthorProfile/AuthorProfile";
 import { useDispatch } from "react-redux";
@@ -14,6 +10,7 @@ import { follow, unfollow } from "../../components/actions/actions";
 import FollowingCard from "../../components/FollowingCard/FollowingCard";
 import NotificationsCard from "../../components/NotificationsCard/NotificationsCard";
 import ButtonD from "../../components/buttons/ButtonD/ButtonD";
+import useGetProfileData from "../../components/_hooks/useGetProfileData";
 
 const FadeIn = ({ children }) => {
   const [isVisib, setIsVisib] = React.useState(false);
@@ -58,119 +55,24 @@ const Wrapper = ({ children, author, userId, user, getData, flag }) => {
 };
 
 const ProfileScreen = () => {
-  const { userId } = useParams();
-  const user = useSelector((state) => state.user);
-  const [author, setAuthor] = useState();
-  const [loading, setLoading] = useState(false);
-  const [isFollowing, setIsFollowing] = useState(true);
   const dispatch = useDispatch();
 
-  // base articles
-  const [articles, setArticles] = useState();
-
-  // user's articles, favorites, replies, and following states
-  const [usersArticles, setUsersArticles] = useState();
-  const [favorites, setFavorites] = useState();
-  const [replies, setReplies] = useState();
-  const [following, setFollowing] = useState();
-
-  // set which state to use (articles, favorites, replies, following)
-  const [flag, setFlag] = useState("");
-
-  const search = useLocation().search;
-  const breadcrumb = new URLSearchParams(search).get("breadcrumb");
-
-  useEffect(() => {
-    const asyncCall = async () => {
-      setLoading(true);
-      setFlag(breadcrumb);
-      // get articles written by author
-      await axios
-        .get(
-          `${
-            process.env.REACT_APP_DOMAIN
-          }/api/articles/getArticlesByAuthor?authorId=${
-            userId === user._id ? user._id : userId
-          }`
-        )
-        .then((response) => {
-          setUsersArticles(response.data.reverse());
-          if (breadcrumb === "articles") {
-            setArticles(response.data);
-          }
-        })
-        .catch((error) => console.log(error));
-      // get author info
-      await axios
-        .get(
-          `${process.env.REACT_APP_DOMAIN}/api/users/getUser?id=${
-            userId === user._id ? user._id : userId
-          }`
-        )
-        .then((response) => setAuthor(response.data))
-        .catch((error) => console.log(error));
-      if (userId === user._id) {
-        // get users favorite articles
-        await axios
-          .get(
-            `${process.env.REACT_APP_DOMAIN}/api/users/getUser?id=${user._id}`
-          )
-          .then((result) => {
-            Promise.all(
-              result.data.favorites.map((x) => {
-                return axios.get(
-                  `${process.env.REACT_APP_DOMAIN}/api/articles/getArticle?articleId=${x}`
-                );
-              })
-            )
-              .then((response) => {
-                setFavorites(response.reverse().map((x) => x.data));
-                if (breadcrumb === "favorites") {
-                  setArticles(response.map((x) => x.data));
-                }
-              })
-              .catch((error) => console.log(error));
-          })
-          .catch((error) => console.log(error));
-        // get other users replies from articles written by user
-        await axios
-          .get(
-            `${process.env.REACT_APP_DOMAIN}/api/replies/getRepliesByAuthor`,
-            {
-              headers: { authorization: `Bearer ${user.accessToken}` },
-            }
-          )
-          .then((response) => {
-            setReplies(response.data.reverse());
-            if (breadcrumb === "notifications") {
-              setArticles(response.data);
-            }
-          })
-          .catch((error) => console.log(error));
-        // get following data of all following users
-        Promise.all(
-          user.following.map((x) => {
-            return axios.get(
-              `${process.env.REACT_APP_DOMAIN}/api/users/getUser?id=${x.userId}`
-            );
-          })
-        )
-          .then((response) => {
-            setFollowing(response.map((x) => x.data));
-          })
-          .catch((error) => console.log(error));
-      }
-      setLoading(false);
-    };
-    asyncCall();
-  }, [
+  let {
     userId,
-    breadcrumb,
-    process.env.REACT_APP_DOMAIN,
-    user._id,
-    user.accessToken,
-    user.following,
-  ]);
+    user,
+    author,
+    loading,
+    isFollowing,
+    setIsFollowing,
+    flag,
+    setFlag,
+    usersArticles,
+    favorites,
+    replies,
+    following,
+    articles,
+    setArticles,
+  } = useGetProfileData();
 
   const getData = (crumb) => {
     if (crumb === "articles") {
